fix: move devDeps into deps
