+++ packages/middleware-endpoint-discovery/package.json
@@ -21,7 +21,6 @@
   },
   "license": "Apache-2.0",
   "devDependencies": {
-    "@smithy/node-config-provider": "^2.0.10",
     "@tsconfig/recommended": "1.0.1",
     "concurrently": "7.0.0",
     "downlevel-dts": "0.10.1",
@@ -32,6 +31,7 @@
   "dependencies": {
     "@aws-sdk/endpoint-cache": "*",
     "@aws-sdk/types": "*",
+    "@smithy/node-config-provider": "^2.0.10",
     "@smithy/protocol-http": "^3.0.3",
     "@smithy/types": "^2.3.1",
     "tslib": "^2.5.0"
