+++ packages/rds-signer/package.json
@@ -35,11 +35,11 @@
     "@smithy/node-config-provider": "^2.0.10",
     "@smithy/protocol-http": "^3.0.3",
     "@smithy/signature-v4": "^2.0.0",
+    "@smithy/types": "^2.3.1",
     "tslib": "^2.5.0"
   },
   "devDependencies": {
     "@aws-sdk/types": "*",
-    "@smithy/types": "^2.3.1",
     "@types/node": "^14.14.31",
     "concurrently": "7.0.0",
     "downlevel-dts": "0.10.1",
