+++ packages/signature-v4-crt/package.json
@@ -25,6 +25,7 @@
     "@aws-sdk/signature-v4-multi-region": "*",
     "@smithy/querystring-parser": "^2.0.0",
     "@smithy/signature-v4": "^2.0.0",
+    "@smithy/types": "^2.3.1",
     "@smithy/util-middleware": "^2.0.0",
     "aws-crt": "^1.15.9",
     "tslib": "^2.5.0"
@@ -33,7 +34,6 @@
     "@aws-crypto/sha256-js": "3.0.0",
     "@aws-sdk/types": "*",
     "@smithy/protocol-http": "^3.0.3",
-    "@smithy/types": "^2.3.1",
     "@tsconfig/recommended": "1.0.1",
     "concurrently": "7.0.0",
     "downlevel-dts": "0.10.1",
