+++ packages/util-dns/package.json
@@ -23,10 +23,10 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "tslib": "^2.5.0"
+    "tslib": "^2.5.0",
+    "@aws-sdk/types": "*"
   },
   "devDependencies": {
-    "@aws-sdk/types": "*",
     "@tsconfig/recommended": "1.0.1",
     "@types/node": "^14.14.31",
     "concurrently": "7.0.0",
