+++ packages/util-dynamodb/package.json
@@ -24,13 +24,16 @@
     "tslib": "^2.5.0"
   },
   "devDependencies": {
-    "@aws-sdk/client-dynamodb": "*",
+    "@aws-sdk/client-dynamodb": "^3.0.0",
     "@tsconfig/recommended": "1.0.1",
     "concurrently": "7.0.0",
     "downlevel-dts": "0.10.1",
     "rimraf": "3.0.2",
     "typedoc": "0.23.23",
     "typescript": "~4.9.5"
+  },
+  "peerDependencies": {
+    "@aws-sdk/client-dynamodb": "^3.0.0"
   },
   "engines": {
     "node": ">=14.0.0"
